+++ app/src/editor.html
@@ -154,7 +154,7 @@
           </button>
           <button id="operation-add-box" title="Add box" class="single-choice">
             <inline-svg>icon-box</inline-svg>
-            <div class="keyboard">q</div>
+            <div class="keyboard">3</div>
           </button>
           <button
             id="operation-link"
@@ -162,23 +162,7 @@
             class="single-choice"
           >
             <inline-svg>icon-link</inline-svg>
-            <div class="keyboard">w</div>
-          </button>
-          <button
-            id="operation-set-title"
-            title="Name boxes & lines"
-            class="single-choice"
-          >
-            <inline-svg>icon-text</inline-svg>
-            <div class="keyboard">a</div>
-          </button>
-          <button
-            id="operation-set-color"
-            title="Color boxes & lines"
-            class="single-choice"
-          >
-            <inline-svg>icon-paint</inline-svg>
-            <div class="keyboard">s</div>
+            <div class="keyboard">4</div>
           </button>
           <button
             id="operation-add-list"
@@ -186,7 +170,15 @@
             class="single-choice"
           >
             <inline-svg>icon-list</inline-svg>
-            <div class="keyboard">z</div>
+            <div class="keyboard">5</div>
+          </button>
+          <button
+            id="operation-set-color"
+            title="Color boxes & lines"
+            class="single-choice"
+          >
+            <inline-svg>icon-paint</inline-svg>
+            <div class="keyboard">6</div>
           </button>
           <button
             id="operation-debug"
@@ -318,6 +310,9 @@
           <button title="Duplicate" data-value="duplicate">
             <inline-svg>icon-duplicate</inline-svg>
           </button>
+          <button title="Set title" data-value="set-title">
+            <inline-svg>icon-text</inline-svg>
+          </button>
         </div>
         <div class="separator-camera"></div>
         <div class="operations">
@@ -431,10 +426,6 @@
             <tr>
               <td><inline-svg>icon-eraser</inline-svg></td>
               <td>Remove one or many boxes, lines, etc. from the canvas.</td>
-            </tr>
-            <tr>
-              <td><inline-svg>icon-text</inline-svg></td>
-              <td>Add or modify the title of a box, list, line, etc.</td>
             </tr>
             <tr>
               <td><inline-svg>icon-paint</inline-svg></td>
