+++ app/src/operations/addBox.ts
@@ -157,6 +157,7 @@ const operation: Operation = {
     placeholderVisual.visible = false;
     parentVisual.visible = false;
   },
+  onDoubleTap: () => {},
 };
 
 export default operation;
