+++ app/src/operations/erase.ts
@@ -287,6 +287,7 @@ const operation: Operation = {
   },
   onPointerEnter: () => {},
   onPointerLeave: () => {},
+  onDoubleTap: () => {},
 };
 
 export default operation;
