+++ app/src/operations/paint.ts
@@ -167,6 +167,7 @@ const operation: Operation = {
   },
   onPointerEnter: () => {},
   onPointerLeave: () => {},
+  onDoubleTap: () => {},
 };
 
 export default operation;
